test(connections): migrate slot/access tests off removed PGlite harness
These DB-backed tests were written against the pre-#3528 PGlite test API
(`../database/test-db`, `createTestSchema`, `seedCommonTestFixtures`),
which main removed when it moved storage tests to real Postgres. After
rebasing, the dead imports broke knip, typecheck, and the unit test job.

Convert all five to the real-Postgres harness, matching the established
087–092 / connection.integration.test.ts pattern:
- Rename DB-backed files to *.integration.test.ts so they route to the
  storage-integration workflow and are excluded from the unit job.
- Use connectTestPgDatabase / resetTestPgDatabase /
  seedCommonTestPgFixtures / closeTestPgDatabase.
- Split slot-resolver: pure SlotResolutionCache/SlotUnresolvedError tests
  stay as a unit *.test.ts; the resolveSlot DB tests move to
  slot-resolver.integration.test.ts.
- 097 migration test now asserts the already-migrated schema's behavior
  (default, CHECK, R4 unique index, aggregation XOR, slot uniqueness).
  Drop the backfill and `down` rollback cases: rolling the shared
  integration DB to 096 / dropping the access column would corrupt the
  schema for sibling integration files. Mirrors 087–092 (up-only).

Verified: 31 tests pass against real Postgres; knip, fmt:check, and tsc
all clean.
